Virtio-loopabck adapter multi-sound demo release
This release is based on epsilon branch and introduces a few modifications
in order to request from the adapter to work with a specific virtio-loopback
modules based on its name under /dev/ directory.

This version can be used to run multiple adapter instances working with
different virtio-loopback transport instances and virtio devices.

Signed-off-by: Timos Ampelikiotis <t.ampelikiotis@virtualopensystems.com>